+++ numba/core/errors.py
@@ -746,6 +746,7 @@ def new_error_context(fmt_, *args, **kwargs):
         e.add_context(_format_msg(fmt_, args, kwargs))
         raise
     except AssertionError:
+        # Let assertion error pass through for shorter traceback in debugging
         raise
     except Exception as e:
         newerr = errcls(e).add_context(_format_msg(fmt_, args, kwargs))
