+++ numba/core/inline_closurecall.py
@@ -398,9 +398,6 @@ class InlineWorker(object):
                 new_var = scope.redefine(var.name, loc=var.loc)
                 var_dict[var.name] = new_var
         self.debug_print("var_dict = ", var_dict)
-        #    replace scope in callee blocks
-        for blk in callee_blocks.values():
-            blk.scope = scope
         replace_vars(callee_blocks, var_dict)
         self.debug_print("After local var rename")
         _debug_dump(callee_ir)
