+++ numba/parfors/parfor.py
@@ -75,8 +75,7 @@ from numba.core.ir_utils import (
     set_index_var_of_get_setitem,
     find_potential_aliases,
     replace_var_names,
-    enforce_single_scope,
-    legalize_single_scope,
+    transfer_scope,
 )
 
 from numba.core.analysis import (compute_use_defs, compute_live_map,
@@ -2786,7 +2785,6 @@ class ParforPass(ParforPassStates):
     def run(self):
         """run parfor conversion pass: replace Numpy calls
         with Parfors when possible and optimize the IR."""
-        enforce_single_scope(self.func_ir)
         self._pre_run()
         # run stencil translation to parfor
         if self.options.stencil:
@@ -2878,14 +2876,12 @@ class ParforPass(ParforPassStates):
                             new_block.append(ir_print)
                 block.body = new_block
 
-        enforce_single_scope(self.func_ir)
         if self.func_ir.is_generator:
             fix_generator_types(self.func_ir.generator_info, self.return_type,
                                 self.typemap)
         if sequential_parfor_lowering:
             lower_parfor_sequential(
                 self.typingctx, self.func_ir, self.typemap, self.calltypes, self.metadata)
-            enforce_single_scope(self.func_ir)
         else:
             # prepare for parallel lowering
             # add parfor params to parfors here since lowering is destructive
@@ -2901,7 +2897,6 @@ class ParforPass(ParforPassStates):
             # Validate parameters:
             for p in parfors:
                 p.validate_params(self.typemap)
-            enforce_single_scope(self.func_ir)
 
             if config.DEBUG_ARRAY_OPT_STATS:
                 name = self.func_ir.func_id.func_qualname
@@ -2915,7 +2910,6 @@ class ParforPass(ParforPassStates):
                 else:
                     print('Function {} has no Parfor.'.format(name))
 
-        enforce_single_scope(self.func_ir)
         return
 
     def _find_mask(self, arr_def):
@@ -3262,7 +3256,6 @@ def lower_parfor_sequential(typingctx, func_ir, typemap, calltypes, metadata):
                               ir_utils.find_max_label(func_ir.blocks))
     parfor_found = False
     new_blocks = {}
-    enforce_single_scope(func_ir)
     scope = next(iter(func_ir.blocks.values())).scope
     for (block_label, block) in func_ir.blocks.items():
         block_label, parfor_found = _lower_parfor_sequential_block(
@@ -3270,9 +3263,7 @@ def lower_parfor_sequential(typingctx, func_ir, typemap, calltypes, metadata):
             scope=scope)
         # old block stays either way
         new_blocks[block_label] = block
-        assert legalize_single_scope(new_blocks)
     func_ir.blocks = new_blocks
-    enforce_single_scope(func_ir)
     # rename only if parfor found and replaced (avoid test_flow_control error)
     if parfor_found:
         func_ir.blocks = rename_labels(func_ir.blocks)
@@ -3280,19 +3271,6 @@ def lower_parfor_sequential(typingctx, func_ir, typemap, calltypes, metadata):
     simplify(func_ir, typemap, calltypes, metadata["parfors"])
     dprint_func_ir(func_ir, "after parfor sequential simplify")
 
-
-def _transfer_scope(old_block, new_scope):
-    old_scope = old_block.scope
-    if old_scope is new_scope:
-        return old_block
-    var_dict = {}
-    for var in old_scope.localvars._con.values():
-        new_var = new_scope.redefine(var.name, loc=var.loc)
-        var_dict[var.name] = new_var
-    # replace scope
-    new_block = old_block#.copy()
-    new_block.scope = new_scope
-    return new_block
 
 def _lower_parfor_sequential_block(
         block_label,
@@ -3314,13 +3292,11 @@ def _lower_parfor_sequential_block(
         # previous block jump to parfor init block
         init_label = next_label()
         prev_block.body.append(ir.Jump(init_label, loc))
-        new_blocks[init_label] = _transfer_scope(inst.init_block, scope)
+        new_blocks[init_label] = transfer_scope(inst.init_block, scope)
         new_blocks[block_label] = prev_block
-        assert legalize_single_scope(new_blocks)
         block_label = next_label()
 
         ndims = len(inst.loop_nests)
-
         for i in range(ndims):
             loopnest = inst.loop_nests[i]
             # create range block for loop
@@ -3362,8 +3338,7 @@ def _lower_parfor_sequential_block(
             l, parfor_found = _lower_parfor_sequential_block(
                 l, b, new_blocks, typemap, calltypes, parfor_found,
                 scope=scope)
-
-            new_blocks[l] = _transfer_scope(b, scope)
+            new_blocks[l] = transfer_scope(b, scope)
         i = _find_first_parfor(block.body)
     return block_label, parfor_found
